Start splitting main into smaller files
